+++ examples/hello_world.rs
@@ -2,6 +2,7 @@
 extern crate glium;
 #[macro_use]
 extern crate imgui;
+extern crate imgui_glium_renderer;
 
 use imgui::*;